J2N.csproj: Added direct dependency on System.Runtime.CompilerServices.Unsafe for legacy targets to bump the version to 6.0.0. This allows the use of Unsafe.AddByteOffset(ref T, nuint) by Marvin (matching upstream). This changes the net45 target to use NetFx.System.Memory instead of System.Memory.

+++ src/J2N/Marvin.cs
@@ -58,7 +58,7 @@ namespace J2N
                 // running in a 64-bit process.
 
                 p0 += Unsafe.ReadUnaligned<uint>(ref data);
-                uint nextUInt32 = Unsafe.ReadUnaligned<uint>(ref Unsafe.AddByteOffset(ref data, (IntPtr)4));
+                uint nextUInt32 = Unsafe.ReadUnaligned<uint>(ref Unsafe.AddByteOffset(ref data, 4));
 
                 // One block round for each of the 32-bit integers we just read, 2x rounds total.
 
@@ -73,7 +73,7 @@ namespace J2N
                 // Requires https://github.com/dotnet/runtime/issues/6794 to be addressed first
                 // before we can realize the full benefits of this.
 
-                data = ref Unsafe.AddByteOffset(ref data, (IntPtr)8);
+                data = ref Unsafe.AddByteOffset(ref data, 8);
             } while (--loopCount > 0);
 
             // n.b. We've not been updating the original 'count' parameter, so its actual value is
@@ -109,7 +109,7 @@ namespace J2N
 
             // Read the last 4 bytes of the buffer.
 
-            uint partialResult = Unsafe.ReadUnaligned<uint>(ref Unsafe.Add(ref Unsafe.AddByteOffset(ref data, ((IntPtr)(count & 7))), -4));
+            uint partialResult = Unsafe.ReadUnaligned<uint>(ref Unsafe.Add(ref Unsafe.AddByteOffset(ref data, (nuint)count & 7), -4));
 
             // The 'partialResult' local above contains any data we have yet to read, plus some number
             // of bytes which we've already read from the buffer. An example of this is given below
